+++ src/routes/projects/projects.json
@@ -161,7 +161,7 @@
       "Make",
       "Adobe Xd"
     ],
-    "link": "https://www.boozallen.com/expertise/products/precog.html",
+    "link": "https://www.boozallen.com/expertise/products/precog",
     "deployedWith": ["Docker"]
   },
   {
